KT-1895 Make it possible to pass 1.0 as Float without explicit casting
 #KT-1895 fixed

+++ compiler/frontend/src/org/jetbrains/jet/lang/resolve/constants/CompileTimeConstantResolver.java
@@ -132,7 +132,7 @@ public class CompileTimeConstantResolver {
         }
         else if (JetTypeChecker.INSTANCE.isSubtypeOf(builtIns.getFloatType(), expectedType)) {
             try {
-                return new DoubleValue(Float.parseFloat(text));
+                return new FloatValue(Float.parseFloat(text));
             }
             catch (NumberFormatException e) {
                 return OUT_OF_RANGE;
